Fix compilation warnings in new JSON code; modified project generation script and input to always use '/' in paths and translate as necessary to '\'
--HG--
branch : xqilla

+++ src/functions/FunctionParseJSON.cpp
@@ -282,11 +282,13 @@ Sequence FunctionParseJSON::createSequence(DynamicContext* context, int flags) c
   yajl_parser_config cfg = { 0 };
   yajl_handle yajl = yajl_alloc(&json2xml_callbacks, &cfg, &env);
 
-  yajl_status stat = yajl_parse(yajl, (unsigned char*)utf8.UTF8Form(), strlen(utf8.UTF8Form()));
+  yajl_status stat = yajl_parse(yajl, (unsigned char*)utf8.UTF8Form(),
+                                (unsigned int)strlen(utf8.UTF8Form()));
   if(stat != yajl_status_ok) {
     XMLBuffer buf;
     buf.append(X("JSON "));
-    unsigned char *str = yajl_get_error(yajl, 1, (unsigned char*)utf8.UTF8Form(), strlen(utf8.UTF8Form()));
+    unsigned char *str = yajl_get_error(yajl, 1, (unsigned char*)utf8.UTF8Form(),
+                                        (unsigned int)strlen(utf8.UTF8Form()));
     buf.append(X((char*)str));
     yajl_free_error(str);
     XQThrow(FunctionException, X("FunctionParseJSON::createSequence"), buf.getRawBuffer());
